fix: 기존 유저에게 채팅창 반복적으로 추가 되는 현상 수정
